feat: Run in DB transaction (#46)

+++ src/application/dtos/common.rs
@@ -2,6 +2,6 @@ use serde::Deserialize;
 
 #[derive(Deserialize)]
 pub struct PaginationQueryParams {
-    pub limit: Option<usize>,
-    pub offset: Option<usize>,
+    pub limit: Option<u64>,
+    pub offset: Option<u64>,
 }
